Changed typedefs for int32 to long for better portability

+++ schnitzelcraft.c
@@ -10,8 +10,8 @@ typedef unsigned char uint8_t;
 typedef signed char int8_t;
 typedef unsigned short uint16_t;
 typedef signed short int16_t;
-typedef unsigned int uint32_t;
-typedef signed int int32_t;
+typedef unsigned long uint32_t;
+typedef signed long int32_t;
 
 //#define fopen_s(f, n, m) (((int)((*f = fopen(n, m)), 0))
